AI_UseCase/requirements.txt
streamlit
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
langchain-community
faiss-cpu
pymupdf
python-docx
pytesseract
pillow
requests
python-dotenv
